+++ docs/changelog.md
@@ -15,6 +15,7 @@ Sprint 2 / December 15, 2019
 - added tests for index.tsx.
 - added webpack-dev-server.
 - added start script.
+- added webpack-dev-server config in webpack config, allowing for webpack-dev-server to work and hot reload.
 
 ## Release 0.1.0
 Sprint 1 / December 1, 2019